Speedup recursion checks
